explicitly include v1 module in jwplatform module and move version to avoid circular imports
