+++ setup.cfg
@@ -1,5 +1,6 @@
 [metadata]
 description-file=README.rst
+version = attr: jwplatform.version.__version__
 
 [aliases]
 test=pytest
